NewLore listview works. Also NewLore save works.
