+++ app/src/main/java/com/example/user/larper/NewLoreFragment.java
@@ -99,7 +99,7 @@ public class NewLoreFragment extends ListFragment {
             mListener = (OnNewLoreFragmentListener) context;
         } else {
             throw new RuntimeException(context.toString()
-                    + " must implement OnInitProfileFragmentListener");
+                    + " must implement OnNewLoreFragmentListener");
         }
     }
 
@@ -111,106 +111,5 @@ public class NewLoreFragment extends ListFragment {
 
     public interface OnNewLoreFragmentListener {
         void gotoLoreInterface();
-    }
-
-    private class SkillsAdapter extends ArrayAdapter<Skill>{
-
-        private ArrayList<Skill> skills;
-        private Context context;
-
-        public SkillsAdapter(@NonNull Context context, @LayoutRes int resource,
-                             @IdRes int textViewResourceId, @NonNull ArrayList<Skill> objects) {
-            super(context, resource, textViewResourceId, objects);
-            this.skills = objects;
-            this.context = context;
-        }
-
-        class ViewHolder{
-            protected EditText skillName;
-            protected EditText skillLevel;
-        }
-
-        public ArrayList<Skill> getItems(){
-            return this.skills;
-        }
-
-
-        @Override
-        public View getView(final int position, View convertView, ViewGroup parent) {
-            View v = convertView;
-            final Skill s = this.skills.get(position);
-            final ArrayAdapter arrAdapter = this;
-
-            // reuse views
-            if (v == null) {
-                v = getActivity().getLayoutInflater().inflate(R.layout.profile_skill_row, null);
-                final ViewHolder vh = new ViewHolder();
-                vh.skillName = (EditText)v.findViewById(R.id.skill_name_et);
-                vh.skillLevel = (EditText)v.findViewById(R.id.skill_level_et);
-
-                vh.skillName.addTextChangedListener(new CustomTextWatcher(vh.skillName, s));
-
-                v.setTag(vh);
-                vh.skillName.setTag(s);
-                vh.skillLevel.setTag(s);
-            } else {
-                ViewHolder vh = (ViewHolder)v.getTag();
-                vh.skillName.setTag(s);
-                vh.skillLevel.setTag(s);
-            }
-
-            ViewHolder vh = (ViewHolder)v.getTag();
-            vh.skillName.setText(s.getName());
-            vh.skillLevel.setText(s.getLevel());
-
-            v.findViewById(R.id.skill_add_btn).setVisibility(View.GONE);
-            v.findViewById(R.id.skill_level_et).setVisibility(View.GONE);
-            Button delSkillBtn = ((Button)v.findViewById(R.id.skill_sub_btn));
-            delSkillBtn.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    skills.remove(s);
-                    arrAdapter.notifyDataSetChanged();
-                    //arrAdapter.remove(s);
-
-                }
-            });
-
-            Log.d("TAG", "Position is: " + position);
-            return v;
-        }
-    }
-
-    private class CustomTextWatcher implements TextWatcher {
-
-        private EditText skillName;
-        private Skill s;
-
-        public CustomTextWatcher(EditText e, Skill s) {
-            this.skillName = e;
-            this.s = s;
-        }
-
-        @Override
-        public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-
-        }
-
-        @Override
-        public void onTextChanged(CharSequence s, int start, int before, int count) {
-
-        }
-
-        @Override
-        public void afterTextChanged(Editable arg0) {
-
-            String text = arg0.toString();
-
-            if (text != null && text.length() > 0) {
-                if (this.skillName.getId() == R.id.skill_name_et) {
-                    this.s.setName(text);
-                }
-            }
-        }
     }
 }
