+++ app/src/main/res/layout/fragment_lore.xml
@@ -7,7 +7,8 @@
 
     <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        android:id="@+id/relativeLayout">
 
         <Button
             android:id="@+id/lore_new_btn"
@@ -22,6 +23,7 @@
     <ListView
         android:id="@android:id/list"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"></ListView>
+        android:layout_height="match_parent"
+        android:layout_below="@+id/relativeLayout"></ListView>
 
 </RelativeLayout>
